workload/requirements.txt
locust ~= 1.3.1
pysocks ~= 1.7.1
pandas ~= 1.1.4
plotly-express ~= 0.4.1
jupyter ~= 1.0.0
